+++ CHANGELOG.md
@@ -4,3 +4,8 @@
 
 ## 0.0.2 - 10.11.2020
 * added repository domain
+
+## 0.0.3 - 10.11.2020
+* added example to package
+* updated pubspec.yaml
+* removed print statements